+++ lib/trailblazer/activity/task_wrap.rb
@@ -14,12 +14,14 @@ module Trailblazer
 
       # Compute runtime arguments necessary to execute a taskWrap per task of the activity.
       # This method is the top-level entry, called only once for the entire activity graph.
-      # [:wrap_static] The taskWrap used for the topmost activity/operation.
-      def invoke(activity, args, wrap_runtime: {}, wrap_static: initial_wrap_static, **circuit_options) # FIXME: why do we need this method?
+      # [:container_activity] the top-most "activity". This only has to look like an Activity
+      #   and exposes a #[] interface so [:wrap_static] can be read and it's compatible to {Trace}.
+      #   It is the virtual activity that "hosts" the actual {activity}.
+      def invoke(activity, args, wrap_runtime: {}, container_activity: container_activity_for(activity), **circuit_options)
         circuit_options = circuit_options.merge(
           runner:       TaskWrap::Runner,
           wrap_runtime: wrap_runtime,
-          activity:     container_activity_for(activity, wrap_static), # for Runner. Ideally we'd have a list of all static_wraps here (even nested).
+          activity:     container_activity, # for Runner. Ideally we'd have a list of all static_wraps here (even nested).
         )
 
         # signal, (ctx, flow), circuit_options =
@@ -29,9 +31,8 @@ module Trailblazer
       # {:extension} API
       # Extend the static taskWrap from a macro or DSL call.
       # Gets executed in {Intermediate.call} which also provides {config}.
-
-      def initial_wrap_static(*)
-        Pipeline.new([Pipeline.Row("task_wrap.call_task", TaskWrap.method(:call_task))])
+      def initial_wrap_static
+        INITIAL_WRAP_STATIC
       end
 
       # This is the top-most "activity" that hosts the actual activity being run.
@@ -39,15 +40,16 @@ module Trailblazer
       # access {activity[:wrap_static]} to compile the effective taskWrap.
       #
       # It's also needed in Trace/Introspect and mimicks the host containing the actual activity.
-      def container_activity_for(activity, wrap_static)
+      #
+      # DICUSS: we could cache that on Strategy/Operation level.
+      def container_activity_for(activity, wrap_static: initial_wrap_static)
         {
           wrap_static:  {activity => wrap_static},
         }
       end
+
+      INITIAL_WRAP_STATIC = Pipeline.new([Pipeline.Row("task_wrap.call_task", TaskWrap.method(:call_task))].freeze)
     end # TaskWrap
   end
 end
-require "trailblazer/activity/task_wrap/pipeline"
-require "trailblazer/activity/task_wrap/call_task"
-require "trailblazer/activity/task_wrap/runner"
-require "trailblazer/activity/task_wrap/extension"
+